+++ modules/core-config/modules/cert-manager/variables.tf
@@ -40,11 +40,6 @@ variable "cluster_name" {
   nullable    = false
 }
 
-variable "workload_identity" {
-  description = "If the cluster has workload identity enabled."
-  type        = bool
-}
-
 variable "cluster_oidc_issuer_url" {
   description = "The URL of the cluster OIDC issuer."
   type        = string