+++ modules/core-config/modules/external-dns/local.tf
@@ -70,16 +70,16 @@ locals {
       create = true
       name   = local.private_service_account_name
 
-      labels = var.workload_identity && local.use_aad_workload_identity ? {
+      labels = local.use_aad_workload_identity ? {
         "azure.workload.identity/use" = "true"
       } : {}
 
-      annotations = var.workload_identity && local.use_aad_workload_identity ? {
+      annotations = local.use_aad_workload_identity ? {
         "azure.workload.identity/client-id" = local.enable_private ? module.identity_private[0].id : ""
       } : {}
     }
 
-    podLabels = merge(var.labels, var.workload_identity && local.use_aad_workload_identity ? {} : {
+    podLabels = merge(var.labels, local.use_aad_workload_identity ? {} : {
       aadpodidbinding = local.enable_private ? module.identity_private[0].name : ""
     })
 
@@ -111,16 +111,16 @@ locals {
       create = true
       name   = local.public_service_account_name
 
-      labels = var.workload_identity && local.use_aad_workload_identity ? {
+      labels = local.use_aad_workload_identity ? {
         "azure.workload.identity/use" = "true"
       } : {}
 
-      annotations = var.workload_identity && local.use_aad_workload_identity ? {
+      annotations = local.use_aad_workload_identity ? {
         "azure.workload.identity/client-id" = local.enable_public ? module.identity_public[0].id : ""
       } : {}
     }
 
-    podLabels = merge(var.labels, var.workload_identity && local.use_aad_workload_identity ? {} : {
+    podLabels = merge(var.labels, local.use_aad_workload_identity ? {} : {
       aadpodidbinding = local.enable_public ? module.identity_public[0].name : ""
     })