+++ modules/core-config/modules/kube-prometheus-stack/identity.tf
@@ -4,7 +4,7 @@ module "identity_grafana" {
   location            = var.location
   resource_group_name = var.resource_group_name
 
-  workload_identity = var.workload_identity && local.use_aad_workload_identity
+  workload_identity = local.use_aad_workload_identity
   oidc_issuer_url   = var.cluster_oidc_issuer_url
 
   name      = "${var.cluster_name}-grafana"
@@ -27,7 +27,7 @@ module "identity_thanos" {
   location            = var.location
   resource_group_name = var.resource_group_name
 
-  workload_identity = var.workload_identity && local.use_aad_workload_identity
+  workload_identity = local.use_aad_workload_identity
   oidc_issuer_url   = var.cluster_oidc_issuer_url
 
   name = "${var.cluster_name}-thanos"